# Optimize bitmap function count, use roaring cardinality method