… be/test/exprs/bitmap_function_test.cpp