@yangzhg 2021-11-22
> Do you have benchmark result for this improvement?

attach to the comment